requirements.txt
# ===== Core =====
numpy==1.26.4
pillow==10.4.0
pyyaml==6.0.2
tqdm==4.66.5
scipy==1.11.4
matplotlib==3.9.1

# ===== Vision / IO =====
opencv-python==4.10.0.84
imageio==2.35.1
imageio-ffmpeg==0.5.1
scikit-image==0.24.0

# ===== DL / Training =====
tensorboard==2.17.0
lpips==0.1.4

# ===== (Optional) but useful =====
# h5py==3.11.0
# einops==0.8.0
